LPS-93514 PrefsPropsTemporarySwapper is only used in MailServiceTestUtil, move as inner class
